Add Immutable annotation to avoid unneeded composable recompositions, use float for word typing speed (double is overkill), move input state strings to UI layer

+++ core/analytics/impl/src/main/java/com/dapsoft/wpmcounter/analytics/impl/domain/SpeedCalculator.kt
@@ -11,5 +11,5 @@ import kotlin.time.Duration
  * - Active time is the cumulative time during which typing was detected.
  */
 internal interface SpeedCalculator {
-    fun calculateWordsPerMinute(validWordsCount: Int, activeTime: Duration): Double
+    fun calculateWordsPerMinute(validWordsCount: Int, activeTime: Duration): Float
 }
